Refactored to move _backend subpackage to signal. Renamed modules to better resemble other parts of codebase.
