--- src/ophyd_async/tango/_backend/__init__.py
@@ -1,21 +1,0 @@
-from ._tango_backend import (
-    AttributeProxy,
-    CommandProxy,
-    TangoSignalBackend,
-    ensure_proper_executor,
-    get_dtype_extended,
-    get_python_type,
-    get_tango_trl,
-    get_trl_descriptor,
-)
-
-__all__ = [
-    "AttributeProxy",
-    "CommandProxy",
-    "ensure_proper_executor",
-    "TangoSignalBackend",
-    "get_python_type",
-    "get_dtype_extended",
-    "get_trl_descriptor",
-    "get_tango_trl",
-]
