+++ src/ophyd_async/tango/base_devices/__init__.py
@@ -1,4 +1,4 @@
-from .base_device import TangoDevice
-from .tango_readable import TangoReadable, tango_polling
+from ._base_device import TangoDevice
+from ._tango_readable import TangoReadable, tango_polling
 
 __all__ = ["TangoDevice", "TangoReadable", "tango_polling"]
