+++ src/ophyd_async/tango/demo/_counter.py
