+++ src/ophyd_async/tango/demo/_tango/__init__.py
@@ -1,3 +1,3 @@
-from .servers import DemoCounter, DemoMover
+from ._servers import DemoCounter, DemoMover
 
 __all__ = ["DemoCounter", "DemoMover"]
