+++ src/ophyd_async/tango/signal/__init__.py
@@ -1,4 +1,4 @@
-from .signal import (
+from ._signal import (
     infer_python_type,
     infer_signal_frontend,
     make_backend,
@@ -8,8 +8,26 @@ from .signal import (
     tango_signal_w,
     tango_signal_x,
 )
+from ._tango_transport import (
+    AttributeProxy,
+    CommandProxy,
+    TangoSignalBackend,
+    ensure_proper_executor,
+    get_dtype_extended,
+    get_python_type,
+    get_tango_trl,
+    get_trl_descriptor,
+)
 
 __all__ = (
+    "AttributeProxy",
+    "CommandProxy",
+    "ensure_proper_executor",
+    "TangoSignalBackend",
+    "get_python_type",
+    "get_dtype_extended",
+    "get_trl_descriptor",
+    "get_tango_trl",
     "infer_python_type",
     "infer_signal_frontend",
     "make_backend",
